Added - `clear` command
What
=================
The `clear` command allows a user to clear the current queue, stop the
currently playing song and remove that too. Essentially it allows them
to start a clean slate and add other songs on instead.
I have no idea why this wasn't ever a function in the first place...

Issue ID: N/A